Log improvements

+++ commands/staff/purge.js
@@ -1,6 +1,10 @@
 const { Command } = require('discord.js-commando');
 const { RichEmbed } = require('discord.js');
 
+const Enmap = require("enmap");
+const { commandsRead, messagesRead, translationsDone, settings } = require('../../data/js/enmap.js');
+const defaultSettings = require('../../data/json/default.json');
+
 module.exports = class purgeCommand extends Command {
 	constructor(client) {
 		super(client, {
@@ -18,14 +22,36 @@ module.exports = class purgeCommand extends Command {
 					prompt: 'How many messages would you like to purge? Max: 99',
 					type: 'integer',
 					validate: text => {
-						if (text <= 99 && text > 2) return true;
+						if (text <= 99 && text >= 2) return true;
 						return 'please keep under 99 and over 2 messages!';
 					}
+				},
+				{
+					key: 'confirmation',
+					prompt: 'Are you sure you want to do this? `yes` | `no`',
+					type: 'string',
+					parse: confirmation => confirmation.toLowerCase(),
+					default: ''
 				}
 			]
 		});
 	}
-	run(msg, { number }) {
-		msg.channel.bulkDelete(number + 1)
+	run(msg, { number, confirmation }) {
+		if (confirmation === 'yes') {
+			msg.channel.bulkDelete(number + 1)
+			
+			settings.ensure(msg.guild.id, defaultSettings);
+			settings.fetchEverything();
+			if (settings.get(msg.guild.id, "log") !== 'none') {
+				if (!msg.guild.channels.find(channel => channel.name == settings.get(msg.guild.id, "log"))) return; 
+				
+				const embed = new RichEmbed()
+				.setFooter(new Date().toLocaleDateString("en-US"))
+				.setDescription(`By: **<@${msg.author.id}>**\nIn: **<#${msg.channel.id}>**\nDeleted: **${number}**`)
+				.setAuthor('Message purge', msg.author.displayAvatarURL)
+				.setColor(0xFF0000);
+				msg.guild.channels.find(channel => channel.name == settings.get(msg.guild.id, "log")).send(embed).catch(console.error);
+			}
+		}  else return msg.reply(`Are you sure you want to purge ${number} messags? Re-run the command with ${number} \`yes\``);
 	}
 };
